+++ backend/clickhouse/traces.go
@@ -20,10 +20,9 @@ import (
 )
 
 const TracesTable = "traces"
-const TracesSamplingTable = "traces_sampling"
+const TracesSamplingTable = "traces_sampling_new"
 const TraceKeysTable = "trace_keys"
 const TraceKeyValuesTable = "trace_key_values"
-const TracesByIdTable = "traces_by_id"
 
 var traceKeysToColumns = map[string]string{
 	string(modelInputs.ReservedTraceKeySecureSessionID): "SecureSessionId",
@@ -395,7 +394,7 @@ func extractEvents(result ClickhouseTraceRow) []*modelInputs.TraceEvent {
 	})
 }
 
-func (client *Client) ReadTrace(ctx context.Context, projectID int, traceID string) ([]*modelInputs.Trace, error) {
+func (client *Client) ReadTrace(ctx context.Context, projectID int, traceID string, timestamp time.Time) ([]*modelInputs.Trace, error) {
 	span, ctx := util.StartSpanFromContext(ctx, "clickhouse.ReadTrace", util.Tag("projectID", projectID), util.Tag("traceID", traceID))
 	defer span.Finish()
 
@@ -403,10 +402,13 @@ func (client *Client) ReadTrace(ctx context.Context, projectID int, traceID stri
 	var err error
 	var args []interface{}
 
-	sb.From(TracesByIdTable).
+	sb.From(TracesSamplingTable).
 		Select(selectTraceColumns).
 		Where(sb.Equal("ProjectId", projectID)).
-		Where(sb.Equal("TraceId", traceID))
+		Where(sb.Equal("TraceId", traceID)).
+		Where(fmt.Sprintf("farmHash64(TraceId) = farmHash64(%s)", sb.Var(traceID))).
+		Where(sb.GreaterEqualThan("Timestamp", timestamp.Add(-1*time.Hour))).
+		Where(sb.LessEqualThan("Timestamp", timestamp.Add(time.Hour)))
 
 	sql, args := sb.BuildWithFlavor(sqlbuilder.ClickHouse)
 
@@ -501,8 +503,8 @@ func (client *Client) TracesKeys(ctx context.Context, projectID int, startDate t
 	return traceKeys, nil
 }
 
-func (client *Client) TracesKeyValues(ctx context.Context, projectID int, keyName string, startDate time.Time, endDate time.Time, limit *int) ([]string, error) {
-	return KeyValuesAggregated(ctx, client, TraceKeyValuesTable, projectID, keyName, startDate, endDate, limit)
+func (client *Client) TracesKeyValues(ctx context.Context, projectID int, keyName string, startDate time.Time, endDate time.Time, query *string, limit *int) ([]string, error) {
+	return KeyValuesAggregated(ctx, client, TraceKeyValuesTable, projectID, keyName, startDate, endDate, query, limit)
 }
 
 func TraceMatchesQuery(trace *TraceRow, filters listener.Filters) bool {