+++ e2e/ruby/rails/demo/app/controllers/logs_controller.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class LogsController < ApplicationController
+  def create
+    Highlight.log('info', 'hello, world!', { foo: 'bar' })
+  end
+end