+++ packages/ui/src/components/Table/Body/styles.css.ts
@@ -1,4 +1,0 @@
-import { style } from '@vanilla-extract/css'
-
-// need for row styling
-export const body = style({})