+++ packages/ui/src/components/Table/styles.css.ts
@@ -1,6 +1,14 @@
 import { style } from '@vanilla-extract/css'
 
-// Needed as a selector in other styles
-export const table = style({})
-export const noBorder = style({})
-export const withSearch = style({})
+import { vars } from '@/vars'
+
+export const BORDER = `1px solid ${vars.theme.static.divider.weak}`
+
+export const table = style({
+	border: BORDER,
+	borderRadius: vars.borderRadius['8'],
+})
+
+export const noBorder = style({
+	border: 0,
+})