+++ sdk/client/src/otel/index.ts
@@ -16,7 +16,7 @@ import { FetchInstrumentation } from '@opentelemetry/instrumentation-fetch'
 import { registerInstrumentations } from '@opentelemetry/instrumentation'
 import { XMLHttpRequestInstrumentation } from '@opentelemetry/instrumentation-xml-http-request'
 import { Resource } from '@opentelemetry/resources'
-import { OTLPTraceExporter } from '@opentelemetry/exporter-trace-otlp-http'
+import { OTLPTraceExporterBrowserWithXhrRetry } from './exporter'
 import {
 	SEMRESATTRS_DEPLOYMENT_ENVIRONMENT,
 	SEMRESATTRS_SERVICE_NAME,
@@ -90,7 +90,7 @@ export const setupBrowserTracing = (config: BrowserTracingConfig) => {
 		)
 	}
 
-	const exporter = new OTLPTraceExporter({
+	const exporter = new OTLPTraceExporterBrowserWithXhrRetry({
 		url: endpoint + '/v1/traces',
 		concurrencyLimit: 10,
 		// Using any because we were getting an error importing CompressionAlgorithm
@@ -268,6 +268,14 @@ export const getTracer = () => {
 	return provider.getTracer(BROWSER_TRACER_NAME)
 }
 
+export const getActiveSpan = () => {
+	return api.trace.getActiveSpan()
+}
+
+export const getActiveSpanContext = () => {
+	return api.context.active()
+}
+
 export const shutdown = async () => {
 	if (provider === undefined) {
 		return
@@ -285,7 +293,7 @@ const getSpanName = (
 	let parsedBody
 	const urlObject = new URL(url)
 	const pathname = urlObject.pathname
-	let spanName = `${method} - ${pathname}`
+	let spanName = `${method.toUpperCase()} - ${pathname}`
 
 	try {
 		parsedBody = typeof body === 'string' ? JSON.parse(body) : body