+++ sdk/client/src/utils/sessionStorage/highlightSession.ts
@@ -1,5 +1,5 @@
 import { SESSION_STORAGE_KEYS } from './sessionStorageKeys'
-import { getItem, removeItem, setItem } from '../storage'
+import { cookieStorage, getItem, removeItem, setItem } from '../storage'
 import { SESSION_PUSH_THRESHOLD } from '../../constants/sessions'
 
 export type SessionData = {
@@ -16,12 +16,18 @@ const getSessionDataKey = (sessionID: string): string => {
 	return `${SESSION_STORAGE_KEYS.SESSION_DATA}_${sessionID}`
 }
 
-export const getSessionSecureID = (): string => {
+let sessionSecureID: string = ''
+
+export const getSessionSecureID = (props?: { local?: true }): string => {
+	if (props?.local) {
+		return sessionSecureID
+	}
 	return getItem(SESSION_STORAGE_KEYS.SESSION_ID) ?? ''
 }
 
 export const setSessionSecureID = (secureID: string) => {
-	setItem(SESSION_STORAGE_KEYS.SESSION_ID, secureID)
+	sessionSecureID = secureID
+	setItem(SESSION_STORAGE_KEYS.SESSION_ID, sessionSecureID)
 }
 
 const getSessionData = (sessionID: string): SessionData | undefined => {
@@ -52,4 +58,16 @@ export const setSessionData = function (sessionData?: SessionData) {
 	if (!sessionData?.sessionSecureID) return
 	const secureID = sessionData.sessionSecureID!
 	setItem(getSessionDataKey(secureID), JSON.stringify(sessionData))
+}
+
+export const loadCookieSessionData = function () {
+	const sessionSecureID = cookieStorage.getItem(
+		SESSION_STORAGE_KEYS.SESSION_ID,
+	)
+	setSessionSecureID(sessionSecureID)
+	const sessionDataKey = getSessionDataKey(sessionSecureID)
+	const sessionDataStr = cookieStorage.getItem(sessionDataKey)
+	try {
+		setSessionData(JSON.parse(sessionDataStr) as SessionData)
+	} catch (e) {}
 }