+++ sdk/firstload/src/index.tsx
@@ -27,16 +27,14 @@ import configureElectronHighlight from './environments/electron.js'
 import firstloadVersion from './__generated/version.js'
 import {
 	getPreviousSessionData,
-	type SessionData,
-	setSessionData,
-	setSessionSecureID,
+	loadCookieSessionData,
 } from '@highlight-run/client/src/utils/sessionStorage/highlightSession.js'
 import { initializeFetchListener } from './listeners/fetch'
 import { initializeWebSocketListener } from './listeners/web-socket'
 import { listenToChromeExtensionMessage } from './browserExtension/extensionListener.js'
-import { setItem } from '@highlight-run/client/src/utils/storage.js'
 import { ErrorMessageType } from '@highlight-run/client/src/types/shared-types'
 import type { Context, Span, SpanOptions, Tracer } from '@opentelemetry/api'
+import { setCookieWriteEnabled } from '@highlight-run/client/src/utils/storage'
 
 enum MetricCategory {
 	Device = 'Device',
@@ -98,6 +96,12 @@ const H: HighlightPublicInterface = {
 				return
 			}
 
+			if (!options?.skipCookieSessionDataLoad) {
+				loadCookieSessionData()
+			} else {
+				setCookieWriteEnabled(false)
+			}
+
 			let previousSession = getPreviousSessionData()
 			sessionSecureID = GenerateSecureID()
 			if (previousSession?.sessionSecureID) {
@@ -110,9 +114,15 @@ const H: HighlightPublicInterface = {
 			}
 			init_called = true
 
-			if (options?.enableOtelTracing) {
-				import('@highlight-run/client/src/otel').then(
-					({ setupBrowserTracing, getTracer: otelGetTracer }) => {
+			initializeFetchListener()
+			initializeWebSocketListener()
+			import('@highlight-run/client/src').then(
+				async ({
+					Highlight,
+					setupBrowserTracing,
+					getTracer: otelGetTracer,
+				}) => {
+					if (options?.enableOtelTracing) {
 						setupBrowserTracing({
 							endpoint: options?.otlpEndpoint,
 							projectId: projectID,
@@ -127,53 +137,27 @@ const H: HighlightPublicInterface = {
 								options?.serviceName ?? 'highlight-browser',
 						})
 						getTracer = otelGetTracer
-					},
-				)
-			}
+					}
 
-			initializeFetchListener()
-			initializeWebSocketListener()
-			import('@highlight-run/client/src').then(async ({ Highlight }) => {
-				highlight_obj = new Highlight(
-					client_options,
-					first_load_listeners,
-				)
-				initializeFetchListener()
-				initializeWebSocketListener()
-				if (!options?.manualStart) {
-					await highlight_obj.initialize()
-				}
-			})
+					highlight_obj = new Highlight(
+						client_options,
+						first_load_listeners,
+					)
+					initializeFetchListener()
+					initializeWebSocketListener()
+					if (!options?.manualStart) {
+						await highlight_obj.initialize()
+					}
+				},
+			)
 
 			const client_options: HighlightClassOptions = {
+				...options,
 				organizationID: projectID,
-				debug: options?.debug,
-				backendUrl: options?.backendUrl,
-				tracingOrigins: options?.tracingOrigins,
-				disableNetworkRecording: options?.disableNetworkRecording,
-				networkRecording: options?.networkRecording,
-				disableBackgroundRecording: options?.disableBackgroundRecording,
-				disableConsoleRecording: options?.disableConsoleRecording,
-				disableSessionRecording: options?.disableSessionRecording,
-				reportConsoleErrors: options?.reportConsoleErrors,
-				consoleMethodsToRecord: options?.consoleMethodsToRecord,
-				privacySetting: options?.privacySetting,
-				enableSegmentIntegration: options?.enableSegmentIntegration,
-				enableCanvasRecording: options?.enableCanvasRecording,
-				enablePerformanceRecording: options?.enablePerformanceRecording,
-				enablePromisePatch: options?.enablePromisePatch,
-				samplingStrategy: options?.samplingStrategy,
-				inlineImages: options?.inlineImages,
-				inlineStylesheet: options?.inlineStylesheet,
-				recordCrossOriginIframe: options?.recordCrossOriginIframe,
 				firstloadVersion,
 				environment: options?.environment || 'production',
 				appVersion: options?.version,
-				serviceName: options?.serviceName,
-				sessionShortcut: options?.sessionShortcut,
 				sessionSecureID: sessionSecureID,
-				storageMode: options?.storageMode,
-				sendMode: options?.sendMode,
 			}
 			first_load_listeners = new FirstLoadListeners(client_options)
 			if (!options?.manualStart) {