+++ sdk/highlight-next/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @highlight-run/next
 
+## 7.5.20
+
+### Patch Changes
+
+-   Updated dependencies [f7fb74a44]
+    -   highlight.run@9.2.2
+    -   @highlight-run/node@3.9.0
+    -   @highlight-run/react@4.0.2
+
+## 7.5.19
+
+### Patch Changes
+
+-   Updated dependencies [0067ea6b5]
+    -   highlight.run@9.2.1
+    -   @highlight-run/node@3.9.0
+    -   @highlight-run/react@4.0.1
+
+## 7.5.18
+
+### Patch Changes
+
+-   Updated dependencies [d67bd4425]
+    -   highlight.run@9.2.0
+    -   @highlight-run/node@3.9.0
+    -   @highlight-run/react@4.0.0
+
 ## 7.5.17
 
 ### Patch Changes